Move config.h out of GUI directory
