+++ Source/CLI/Help.cpp
@@ -12,7 +12,7 @@
 
 //---------------------------------------------------------------------------
 #include "CLI/Help.h"
-#include "GUI/Qt/config.h"
+#include "Common/config.h"
 //---------------------------------------------------------------------------
 
 //***************************************************************************
